training/requirements.txt
# nordocr training dependencies
# Install: pip install -r requirements.txt

# Core (January 2026)
torch>=2.10.0
torchvision>=0.25.0
timm>=1.0.24

# PARSeq model (install from repo after cloning)
# pip install -e ./parseq

# Data loading & augmentation
Pillow>=12.1.0
lmdb>=1.7.0
albumentations>=2.0.8

# Training utilities
pytorch-lightning>=2.6.0
torchmetrics>=1.8.0

# ONNX export & TRT engine build
onnx>=1.20.0
onnxsim>=0.4.36
onnxruntime>=1.24.0
onnxoptimizer>=0.3.13

# Monitoring
tensorboard>=2.18.0
tqdm>=4.66.0

# Config & utilities
pyyaml>=6.0
numpy>=2.0.0
pandas>=2.2.0
